Use plain datetime.now() in test
datetime.datetime.utcnow() is deprecated because it is too easy of a
footgun.

Since we don't need timezone-aware objects, now() has the same behaviour
(to the difference of the local time difference to UTC).

The exact time used (or timezone-equivalent) has no impact on this test,
the only requirements is for both calls to have the same timestamp.